Moved helper into its own file
